payment implement

+++ src/main/java/vn/framgia/dao/IBookingDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import vn.framgia.model.Booking;
 
-public interface IBookingDAO extends IGenericDAO<Booking, Long>{
+public interface IBookingDAO extends IGenericDAO<Booking, Integer>{
 	public List<Booking> findAllOrderClientId();
 	public List<Booking> findBillByNameClient(String nameClient);
 }
